+++ src/main/application.js
@@ -1,10 +1,10 @@
 import { app } from 'electron'
 import fs from 'fs'
 import path from 'path'
-import Window from './window'
+import ApplicationWindow from './application-window'
 import * as HostsFileManager from '../renderer/utils/hosts-file-manager'
 
-const WINDOW_INIT_FILE = path.join(app.getPath('userData'), 'window.json')
+const PATH_WINDOW_OPTIONS = path.join(app.getPath('userData'), 'window.json')
 
 export default class Application {
   load() {
@@ -15,24 +15,24 @@ export default class Application {
       return
     }
 
-    const settings = this.loadWindowSettings()
+    const options = this.loadWindowOptions()
 
-    this.window = new Window(this)
-    this.window.open(settings)
+    this.window = new ApplicationWindow(this)
+    this.window.open(options)
   }
   removeWindow() {
     this.window = null
   }
-  loadWindowSettings() {
+  loadWindowOptions() {
     try {
-      return JSON.parse(fs.readFileSync(WINDOW_INIT_FILE))
+      return JSON.parse(fs.readFileSync(PATH_WINDOW_OPTIONS))
     } catch (e) {
       return null
     }
   }
-  saveWindowSettings(settings) {
+  saveWindowOptions(options) {
     try {
-      fs.writeFileSync(WINDOW_INIT_FILE, JSON.stringify(settings))
+      fs.writeFileSync(PATH_WINDOW_OPTIONS, JSON.stringify(options))
     } catch (e) {
       //
     }
